♻️ Refactor: 상품 서비스 로직 개선 및 불필요한 파일 삭제 #43
