+++ src/components/product/ProductForm.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Button from '../common/Button';
 import ProductMapModal from './ProductMapModal';
-import Label from '../ui/Label';
+import Label from '../common/Label';
 import Input from '../common/Input';
 import Select from '../common/Select';
 import TextArea from '../common/TextArea';
